+++ src/snowflake/snowpark/session.py
@@ -16,6 +16,7 @@ import tempfile
 import warnings
 from array import array
 from functools import reduce
+from importlib.metadata import PackageNotFoundError, distribution, distributions
 from logging import getLogger
 from threading import RLock
 from types import ModuleType
@@ -33,7 +34,7 @@ from typing import (
 )
 
 import cloudpickle
-import pkg_resources
+from packaging.requirements import Requirement
 
 from snowflake.connector import ProgrammingError, SnowflakeConnection
 from snowflake.connector.options import installed_pandas, pandas
@@ -64,6 +65,7 @@ from snowflake.snowpark._internal.packaging_utils import (
     DEFAULT_PACKAGES,
     ENVIRONMENT_METADATA_FILE_NAME,
     IMPLICIT_ZIP_FILE_NAME,
+    contains_version,
     delete_files_belonging_to_packages,
     detect_native_dependencies,
     get_signature,
@@ -1249,7 +1251,7 @@ class Session:
             >>> len(session.get_packages())
             0
         """
-        package_name = pkg_resources.Requirement.parse(package).key
+        package_name = Requirement(package).name
         if package_name in self._packages:
             self._packages.pop(package_name)
         else:
@@ -1369,64 +1371,49 @@ class Session:
         ignore_packages = {} if ignore_packages is None else ignore_packages
 
         packages = []
-        for package in pkg_resources.working_set:
-            if package.key in ignore_packages:
-                _logger.info(f"{package.key} found in environment, ignoring...")
+        for package in distributions():
+            if package.name in ignore_packages:
+                _logger.info(f"{package.name} found in environment, ignoring...")
                 continue
-            if package.key in DEFAULT_PACKAGES:
-                _logger.info(f"{package.key} is available by default, ignoring...")
+            if package.name in DEFAULT_PACKAGES:
+                _logger.info(f"{package.name} is availna meble by default, ignoring...")
                 continue
             version_text = (
-                "==" + package.version if package.has_version() and not relax else ""
+                "==" + package.version if package.version and not relax else ""
             )
-            packages.append(f"{package.key}{version_text}")
+            packages.append(f"{package.name}{version_text}")
 
         self.add_packages(packages)
 
     @staticmethod
     def _parse_packages(
         packages: List[Union[str, ModuleType]]
-    ) -> Dict[str, Tuple[str, bool, pkg_resources.Requirement]]:
+    ) -> Dict[str, Tuple[str, bool, Requirement]]:
         package_dict = dict()
         for package in packages:
             if isinstance(package, ModuleType):
                 package_name = MODULE_NAME_TO_PACKAGE_NAME_MAP.get(
                     package.__name__, package.__name__
                 )
-                package = f"{package_name}=={pkg_resources.get_distribution(package_name).version}"
+                package = f"{package_name}=={distribution(package_name).version}"
                 use_local_version = True
             else:
                 package = package.strip().lower()
                 if package.startswith("#"):
                     continue
                 use_local_version = False
-            package_req = pkg_resources.Requirement.parse(package)
-            # get the standard package name if there is no underscore
-            # underscores are discouraged in package names, but are still used in Anaconda channel
-            # pkg_resources.Requirement.parse will convert all underscores to dashes
-            # the regexp is to deal with case that "_" is in the package requirement as well as version restrictions
-            # we only extract the valid package name from the string by following:
-            # https://packaging.python.org/en/latest/specifications/name-normalization/
-            # A valid name consists only of ASCII letters and numbers, period, underscore and hyphen.
-            # It must start and end with a letter or number.
-            # however, we don't validate the pkg name as this is done by pkg_resources.Requirement.parse
-            # find the index of the first char which is not an valid package name character
-            package_name = package_req.key
-            if not use_local_version and "_" in package:
-                reg_match = re.search(r"[^0-9a-zA-Z\-_.]", package)
-                package_name = package[: reg_match.start()] if reg_match else package
-
-            package_dict[package] = (package_name, use_local_version, package_req)
+            package_req = Requirement(package)
+            package_dict[package] = (package_req.name, use_local_version, package_req)
         return package_dict
 
     def _get_dependency_packages(
         self,
-        package_dict: Dict[str, Tuple[str, bool, pkg_resources.Requirement]],
+        package_dict: Dict[str, Tuple[str, bool, Requirement]],
         validate_package: bool,
         package_table: str,
         current_packages: Dict[str, str],
         statement_params: Optional[Dict[str, str]] = None,
-    ) -> List[pkg_resources.Requirement]:
+    ) -> List[Requirement]:
         # Keep track of any package errors
         errors = []
 
@@ -1442,16 +1429,19 @@ class Session:
         unsupported_packages: List[str] = []
         for package, package_info in package_dict.items():
             package_name, use_local_version, package_req = package_info
-            package_version_req = package_req.specs[0][1] if package_req.specs else None
+            package_specifier = package_req.specifier if package_req.specifier else None
 
             if validate_package:
                 if package_name not in valid_packages or (
-                    package_version_req
-                    and not any(v in package_req for v in valid_packages[package_name])
+                    package_specifier
+                    and not any(
+                        contains_version(package_specifier, v)
+                        for v in valid_packages[package_name]
+                    )
                 ):
                     version_text = (
-                        f"(version {package_version_req})"
-                        if package_version_req is not None
+                        f"(version {package_specifier})"
+                        if package_specifier is not None
                         else ""
                     )
                     if is_in_stored_procedure():  # pragma: no cover
@@ -1491,9 +1481,7 @@ class Session:
                     continue
                 elif not use_local_version:
                     try:
-                        package_client_version = pkg_resources.get_distribution(
-                            package_name
-                        ).version
+                        package_client_version = distribution(package_name).version
                         if package_client_version not in valid_packages[package_name]:
                             _logger.warning(
                                 f"The version of package '{package_name}' in the local environment is "
@@ -1501,7 +1489,7 @@ class Session:
                                 f"requirement '{package}'. Your UDF might not work when the package version "
                                 f"is different between the server and your local environment."
                             )
-                    except pkg_resources.DistributionNotFound:
+                    except PackageNotFoundError:
                         _logger.warning(
                             f"Package '{package_name}' is not installed in the local environment. "
                             f"Your UDF might not work when the package is installed on the server "
@@ -1531,7 +1519,7 @@ class Session:
         elif len(errors) > 0:
             raise RuntimeError(errors)
 
-        dependency_packages: List[pkg_resources.Requirement] = []
+        dependency_packages: List[Requirement] = []
         if len(unsupported_packages) != 0:
             _logger.warning(
                 f"The following packages are not available in Snowflake: {unsupported_packages}."
@@ -1654,14 +1642,14 @@ class Session:
         # Add dependency packages
         for package in dependency_packages:
             name = package.name
-            version = package.specs[0][1] if package.specs else None
+            version = package.specifier if package.specifier else None
 
             if name in result_dict:
                 if version is not None:
                     added_package_has_version = "==" in result_dict[name]
                     if added_package_has_version and result_dict[name] != str(package):
                         raise ValueError(
-                            f"Cannot add dependency package '{name}=={version}' "
+                            f"Cannot add dependency package '{name}{version}' "
                             f"because {result_dict[name]} is already added."
                         )
                     result_dict[name] = str(package)
@@ -1685,7 +1673,7 @@ class Session:
         package_table: str,
         package_dict: Dict[str, str],
         custom_package_usage_config: Dict[str, Any],
-    ) -> List[pkg_resources.Requirement]:
+    ) -> List[Requirement]:
         """
         Uploads a list of Pypi packages, which are unavailable in Snowflake, to session stage.
 
@@ -1696,7 +1684,7 @@ class Session:
                 been added explicitly so far using add_packages() or other such methods.
 
         Returns:
-            List[pkg_resources.Requirement]: List of package dependencies (present in Snowflake) that would need to be added
+            List[packaging.requirements.Requirement]: List of package dependencies (present in Snowflake) that would need to be added
             to the package dictionary.
 
         Raises:
@@ -1843,7 +1831,7 @@ class Session:
 
     def _load_unsupported_packages_from_stage(
         self, environment_signature: str, cache_path: str
-    ) -> List[pkg_resources.Requirement]:
+    ) -> List[Requirement]:
         """
         Uses specified stage path to auto-import a group of unsupported packages, along with its dependencies. This
         saves time spent on pip install, native package detection and zip upload to stage.
@@ -1864,7 +1852,7 @@ class Session:
             environment_signature (str): Unique hash signature for a set of unsupported packages, computed by hashing
             a sorted tuple of unsupported package requirements (package versioning included).
         Returns:
-            Optional[List[pkg_resources.Requirement]]: A list of package dependencies for the set of unsupported packages requested.
+            Optional[List[packaging.requirements.Requirement]]: A list of package dependencies for the set of unsupported packages requested.
         """
         # Ensure that metadata file exists
         metadata_file = f"{ENVIRONMENT_METADATA_FILE_NAME}.txt"
@@ -1897,8 +1885,7 @@ class Session:
         }
 
         dependency_packages = [
-            pkg_resources.Requirement.parse(package)
-            for package in metadata[environment_signature]
+            Requirement(package) for package in metadata[environment_signature]
         ]
         _logger.info(
             f"Loading dependency packages list - {metadata[environment_signature]}."