+++ CHANGELOG.md
@@ -26,6 +26,8 @@
 - Added support for 'Service' domain to `session.lineage.trace` API. 
 
 #### Improvements
+- Removed the use of `pkg_resources`, which is deprecated. No more deprecation warnings about it. Used `packaging` and `importlib.metadata` to replace `pkg_resources`.
+
 
 #### Bug Fixes